feat: Post 리스트를 PostPreviewResponse 리스트로 매핑하는 메소드 추가

+++ backend/src/main/java/our/portfolio/devspace/domain/post/dto/PostPreviewResponse.java
@@ -23,7 +23,7 @@ public class PostPreviewResponse {
     private final int commentCount;
 
     @Builder
-    public PostPreviewResponse(int id, SimpleProfileResponse profile, LocalDateTime createdDate, List<String> hashtags, String title, String content, int likeCount,
+    public PostPreviewResponse(long id, SimpleProfileResponse profile, LocalDateTime createdDate, List<String> hashtags, String title, String content, int likeCount,
         int commentCount) {
         this.id = id;
         this.profile = profile;
